Caesar-Cipher for Python
An attempt to create a script that encrypt a message in the Caesar-cipher way. It's very simple, and there are some points that could be better designed, like the alphabet list, that is just  the alphabet written 2 times, where it should have an method to make the list go to the start after reaching the letter z, and some other points, that i will be upgrading over the time.